Crate mint_module_tests

Enums

Cmd 🔒

Functions

main 🔒

restore 🔒

sanity 🔒

test_note_consoliation 🔒

Test note consolidation, which at the time of writing basically means that once client accumulates too many notes of certain denomination, any transaction building will include excessive notes as extra inputs, to consolidate them into higher denominations.

test_restore_gap_test