mint_module_tests

Function test_note_consoliation

pub(crate) async fn test_note_consoliation(fed: &Federation) -> Result<()>

Test note consolidation, which at the time of writing basically means that once client accumulates too many notes of certain denomination, any transaction building will include excessive notes as extra inputs, to consolidate them into higher denominations.

In the future we will probably change the whole thing and delete this thing.